requirements.txt
Django==1.4.1
PIL==1.1.7
South==0.7.6
argparse==1.2.1
distribute==0.6.24
django-debug-toolbar==0.9.4
django-filebrowser==3.5.0
django-grappelli==2.4.0
django-registration==0.8
psycopg2==2.4.5
gevent==0.13.8
gunicorn==0.14.6
python-dateutil==2.1
six==1.2.0
vobject==0.8.1c
wsgiref==0.1.2
xlwt==0.7.4
